+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added examples and tests for `terra-table` to demonstrate auto layout.
+
 ## 1.78.0 - (March 29, 2024)
 
 * Changed