+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Removed `sortIndicator` from the first column of Table With Sections example.
+
 ## 1.78.0 - (March 29, 2024)
 
 * Changed